lsusb: make -t option work if proc file is not present.
This patch comes from the suse usbutils tree.  It implements a
whole new implementation for 'lsusb -t' if usbfs is not mounted.

It's not the best implementation, but it gets things working for
now, which is a good thing to have while we work on making
the package work properly with no usbfs mounted.

Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>