usbutils.pc: use datarootdir
If you want to configure paths so that the usb.ids is stored in a proper
subdir of the /usr/share tree, you currently then have to manually move
the .pc file back up to /usr/share.  If we based the pkgconfig dir on the
datarootdir variable however, we can easily decouple the install of the
usb.ids file file from the rest of the files (man/etc...).

Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>