drop dependency on usb.ids
Also remove usb.ids from the repository. [Note that these were probably
never used by distributions regarless, as most distros ship the usb.ids
directly from upstream.]

Hardcode the usb-spec information that used to be in usb.ids,
but which was not moved to hwdb.

Increase the size of tha hash table from 16 to 512, though using a
hash at all is probably overkill at this point.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>