lsusb: detach the kernel driver before claiming the interface
To display a report descriptor for HID device, we have to success claim
the interface. So, we need to detach the kernel driver before claming the
interface.

The libusb_set_auto_detach_kernel_driver() in libusb automatically do detach
the kernel driver on an interface when claiming the interface, and attach it
when releasing the interface.

Signed-off-by: Jaejoong Kim <climbbb.kim@gmail.com>