sysfs: Don't return bogus data for devices under a hub
commit dddb696e0172 ("lsusb: Read unkown names from sysfs device desc.")
added code to query a device's self-reported vendor and product names
from sysfs in the case where no names were present in the udev hwdb.
However, as far as I can tell that code was and still is totally broken
for any device that isn't directly under a root hub: this is because it
uses the libusb_get_port_number() function, which only returns the leaf
port number, but treats the returned number as if it's a full device
path.

For example, when trying to query sysfs for the device 2-8.3 (a device
on Port 3 of a hub that's on Port 8 of Bus 2), it will instead
erroneously query device 2-3, ignoring the first hub altogether. This
results in nonsensical output from lsusb, such as a device identified as
a "Bose Corp. Steam Controller".

To fix the issue, use the libusb_get_port_numbers() function to get a
full list of port numbers from the root and use that to construct the
proper dotted devpath. Additionally, special case root devices, which
have the sysfs name "usbN", where "N" is the bus number. Thanks to Tian
Yunhao, who submitted a different fix for this same issue, for spotting
that last piece.

Fixes: dddb696e0172 ("lsusb: Read unkown names from sysfs device desc.")
Signed-off-by: Thomas Hebb <tommyhebb@gmail.com>