LICENSES: add symlink to handle "or-later" issue for GPL-2.0
reuse wants to see license files that match the SPDX tag, but for the
"fun" that the FSF seems to have inflicted on the "or-later" moniker for
the GPL licenses, having a duplicated copy of the file is pointless.

So just add a symlink to point to the same thing and all is good as far
as `reuse lint` cares.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>