lsusb: don't complain on EAGAIN
Given the requested decriptors are optional, seems like lsusb
should NOT use perror() to communicate why it can't get
the descriptors in question (output is directed to stderr, not stdout).

Further, if we were told to "come back later" (EGAIN),
then it's really not an error.

Only emit this output when user adds "-vv" to the command line.

Signed-off-by: Grant Grundler <grundler@chromium.org>