usb-devices: use 'local' variable type to handle recursion
When recursing into a long USB tree, the local variables in the
print_device() function would get confused and take on the value of the
previous device it printed.  This caused devices to not get printed out
at all, the exact opposite of what we wanted.

Resolve this by using the non-POSIX 'local' variable declaration.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>