lsusb.py.in: add another default path for usb.ids
Most distros put usb.ids in /usr/share/hwdata/ now, so add that as a
default search path to make working with the script directly out of the
source directory easier.  This does not change the ability for the build
system to add their own default path, so distros should not have any
functional change at all.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>