lsusb: fix up [unknown] vendor and product strings.
In commit 8a80f70a8afb ("names.c: if a string can not be found in the
usb.ids file, return [unknown]"), the logic for determining the vendor
and product for the USB device was changed to ONLY rely on the usb.ids
file, instead of falling back to any vendor/product strings that are on
the device itself.

This had the bad side-affect of causing many devices that previously
showed vendor/product information just fine, to now show up as
[unknown].

Fix this up to fall back to looking at the string in the device itself,
and ONLY if the string is not present, will '[unknown]' be printed out.

This changes the output on one of my machines from:
	Bus 003 Device 002: ID 32ac:0002 [unknown] [unknown]
to:
	Bus 003 Device 002: ID 32ac:0002 Framework HDMI Expansion Card

This should resolve github issue #177, many thanks to the people
reporting it quickly there.

Fixes: 8a80f70a8afb ("names.c: if a string can not be found in the usb.ids file, return [unknown]")
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>