Include "negotiated speed" in device dump
It used to be simple to make sure a device connection was using
USB 3.0+ speed. Check the Type-A connector and make sure it is
blue.

The color code was dropped with USB-C cables. Some of them only
have USB-2 wires, which made some devices silently fallback on
USB-2 speed.

This patch displays the "negotiated speed" of a device so users
can make sure the device is working at the expected speed.

Signed-off-by: Fabien Sanglard <fabien.sanglard@gmail.com>