usbreset: replace some unbounded strcpy() calls
We "know" the sysfs attribute is not going to be bigger than the buffer,
but the thousands of "ooh, a strcpy() call is bad for you!" scanning
tools do not.  So to shut them up, use strncpy() just to make things
quiet and for us to stop getting foolish reports.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>