lsusb: turn a FIXME comment for dump_comm_descriptor() into something real
In dump_comm_descriptor() we have a FIXME that we aren't going to deal
with here anytime soon.  If people show up with a device that has
additional descriptors, we'll be glad to add them.  But until that
happens, just comment it as such as we aren't going to change the code
any time soon.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>