usb-devices: Make devcount 'local' to handle recursion
Commit b1c31712134d2b28877b (usb-devices: use 'local' variable type
to handle recursion) made several variables local to print_device()
to handle recursion. However, devcount was not defined as a local
variable. It is currently treated as a global variable causing it to
retain values updated across function calls. As a result, USB devices
with the same busnum also have the same 'cnt'.

Define devcount as a local variable so 'cnt' is correctly reported for
USB devices sharing the same bus and in the same level.

Signed-off-by: Sahil Siddiq <sahilcdq@proton.me>