Fix display of HID descriptors
The switch from hash tables to arrays + linear lookups created a bug --
`names_genericstrtable` uses two variables (`h` and `t`) inconsistently
in a loop, causing it to always return `NULL`, except for the first
element in the array table.

This in turn causes everything (except "Usage Page") in the HID
descriptors to be `(null)` instead of the right string.

Signed-off-by: Waffle Lapkin <waffle.lapkin@gmail.com>