usb-devices: Make devnum 'local' to handle recursion
Commit b1c3171 (usb-devices: use 'local' variable type to handle
recursion) made several variables local to print_device() to handle
recursion. However, devnum was not defined as a local variable. It is
currently treated as a global variable causing it to retain values
updated across function calls.

Since, a device uses its parent device's devnum to report its prnt,
the prnt values of devices which share the same parent are incorrect.

Define devnum as a local variable so 'prnt' is correctly reported for
USB devices sharing the same parent.

Signed-off-by: Sahil Siddiq <sahilcdq0@gmail.com>